requirements.txt
playwright
python-dotenv
customtkinter
pytest
pytest-mock